refactor(components): extract control primitives (Phase 6 #4)
Splits Button, Checkbox, Select, and Slider out of builtins.tsx into
src/components/primitives/controls.tsx. builtins.tsx re-exports them
and drops the now-unused resolvePointer import.

No logic changes. Refs #43.
